requirements/base.txt
Django
psycopg2
python-decouple
django-split-settings
djangorestframework
gevent
gunicorn
